requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_py-xgboost-mutex=2.0=cpu_0
alabaster=0.7.12=py37_0
alembic=1.0.9=py37_0
appdirs=1.4.4=pyh9f0ad1d_0
argon2-cffi=20.1.0=py37h4ab8f01_2
asn1crypto=1.4.0=pyh9f0ad1d_0
async_generator=1.10=py_0
attrs=20.2.0=pyh9f0ad1d_0
azure-core=1.8.2=pyh9f0ad1d_0
azure-storage-blob=12.5.0=pyh9f0ad1d_0
babel=2.8.0=py_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.1=py_0
blas=1.0=mkl
bleach=3.2.1=pyh9f0ad1d_0
blinker=1.4=py_1
blosc=1.20.1=ha925a31_0
brotli=1.0.9=ha925a31_2
brotlipy=0.7.0=py37h0013d47_1001
bzip2=1.0.8=he774522_3
ca-certificates=2020.10.14=0
certifi=2020.6.20=py37_0
cffi=1.14.3=py37hd6b71e5_1
chardet=3.0.4=py37hf50a25e_1008
charls=2.1.0=h33f27b4_2
click=7.1.2=pyh9f0ad1d_0
cloudpickle=1.6.0=py_0
colorama=0.4.4=pyh9f0ad1d_0
configparser=5.0.1=py_0
cryptography=3.1.1=py37hd6b71e5_1
cycler=0.10.0=py_2
cytoolz=0.11.0=py37h4ab8f01_1
dask-core=2.30.0=py_0
databricks-cli=0.9.1=py_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
docker-py=4.3.1=py37hc8dfbb8_1
docker-pycreds=0.4.0=py_0
docutils=0.16=py37_1
entrypoints=0.3=py37hc8dfbb8_1002
flask=1.1.2=pyh9f0ad1d_0
freetype=2.10.4=hd328e21_0
giflib=5.2.1=h2fa13f4_2
gitdb=4.0.5=py_0
gitpython=3.1.11=py_0
gorilla=0.3.0=py_0
icc_rt=2019.0.0=h0cc432a_1
icu=67.1=h33f27b4_0
idna=2.10=pyh9f0ad1d_0
imagecodecs=2020.5.30=py37hfde3dd5_4
imageio=2.9.0=py_0
imagesize=1.2.0=py_0
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
intel-openmp=2020.2=254
ipykernel=5.3.4=py37h7b7c402_1
ipython=5.8.0=py37_1
ipython_genutils=0.2.0=py_1
ipywidgets=7.5.1=pyh9f0ad1d_1
isodate=0.6.0=py_1
itsdangerous=1.1.0=py_0
jinja2=2.11.2=pyh9f0ad1d_0
joblib=0.17.0=py_0
jpeg=9d=he774522_0
jsonschema=3.2.0=py_2
jupyter=1.0.0=py_2
jupyter_client=6.1.7=py_0
jupyter_console=5.2.0=py37_1
jupyter_core=4.6.3=py37hc8dfbb8_2
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
jxrlib=1.1=hfa6e2cd_2
kiwisolver=1.2.0=py37h1fb7aa8_1
lcms2=2.11=he1115b7_0
lerc=2.2=ha925a31_0
libaec=1.0.4=he025d50_1
libclang=10.0.1=default_hf44288c_1
libpng=1.6.37=ha81a0f5_2
libprotobuf=3.13.0.1=h200bbdf_0
libsodium=1.0.18=h62dcd97_1
libtiff=4.1.0=h885aae3_6
libwebp-base=1.1.0=hfa6e2cd_3
libxgboost=0.90=1
libzopfli=1.0.3=ha925a31_0
llvmlite=0.34.0=py37h8b575af_2
lz4-c=1.9.2=h62dcd97_2
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
mako=1.1.3=pyh9f0ad1d_0
markupsafe=1.1.1=py37h0013d47_2
matplotlib-base=3.3.2=py37h64d6e1c_1
mistune=0.8.4=py37h4ab8f01_1002
mkl=2020.2=256
mkl-service=2.3.0=py37hb782905_0
mkl_fft=1.2.0=py37h45dec08_0
mkl_random=1.1.1=py37h47e9c7a_0
mlflow=1.11.0=py37hc8dfbb8_1
msrest=0.6.19=pyh9f0ad1d_0
msys2-conda-epoch=20160418=1
nbclient=0.5.1=py_0
nbconvert=6.0.7=py37hc8dfbb8_1
nbformat=5.0.8=py_0
nest-asyncio=1.4.1=py_0
networkx=2.5=py_0
notebook=6.1.4=py37hc8dfbb8_1
numba=0.51.2=py37h414f9d2_0
numpy=1.19.1=py37h5510c5b_0
numpy-base=1.19.1=py37ha3acd2a_0
oauthlib=3.0.1=py_0
olefile=0.46=pyh9f0ad1d_1
openjpeg=2.3.1=h57dd2e7_3
openssl=1.1.1h=he774522_0
packaging=20.4=pyh9f0ad1d_0
pandas=1.1.3=py37h06f6abc_2
pandoc=2.11.0.2=hf4a77e7_0
pandocfilters=1.4.2=py_1
pickleshare=0.7.5=py_1003
pillow=8.0.1=py37hb71ed58_0
pip=20.2.4=py37_0
prometheus_client=0.8.0=pyh9f0ad1d_0
prometheus_flask_exporter=0.18.1=pyh9f0ad1d_0
prompt_toolkit=1.0.15=py_1
protobuf=3.13.0.1=py37h1834ac0_1
py-xgboost=0.90=py37_1
pycparser=2.20=pyh9f0ad1d_2
pygments=2.7.1=py_0
pyjwt=1.7.1=py_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.12.3=py37h1834ac0_4
pyqt5-sip=4.19.18=pypi_0
pyqtchart=5.12=pypi_0
pyqtwebengine=5.12.1=pypi_0
pyrsistent=0.17.3=py37h4ab8f01_1
pysocks=1.7.1=py37hf50a25e_2
python=3.7.9=h60c2a47_0
python-dateutil=2.8.1=py_0
python-editor=1.0.4=py_0
python_abi=3.7=1_cp37m
pytz=2020.1=pyh9f0ad1d_0
pywavelets=1.1.1=py37h3a3b6f7_3
pywin32=227=py37hfa6e2cd_0
pywinpty=0.5.7=py37hc8dfbb8_1
pyyaml=5.3.1=py37h0013d47_1
pyzmq=19.0.2=py37h453f00a_2
qt=5.12.9=hb2cf2c5_0
qtconsole=4.7.7=pyh9f0ad1d_0
qtpy=1.9.0=py_0
querystring_parser=1.2.4=py_0
requests=2.24.0=pyh9f0ad1d_0
requests-oauthlib=1.3.0=pyh9f0ad1d_0
scikit-image=0.17.2=py37h414f9d2_3
scikit-learn=0.23.2=py37h47e9c7a_0
scipy=1.5.2=py37h9439919_0
send2trash=1.5.0=py_0
setuptools=50.3.0=py37h9490d1a_1
shap=0.36.0=py37h414f9d2_0
simplegeneric=0.8.1=py_1
six=1.15.0=py_0
slicer=0.0.3=py_0
smmap=3.0.4=pyh9f0ad1d_0
snappy=1.1.8=ha925a31_3
snowballstemmer=2.0.0=py_0
sphinx=3.2.1=py_0
sphinxcontrib-applehelp=1.0.2=py_0
sphinxcontrib-devhelp=1.0.2=py_0
sphinxcontrib-htmlhelp=1.0.3=py_0
sphinxcontrib-jsmath=1.0.1=py_0
sphinxcontrib-qthelp=1.0.3=py_0
sphinxcontrib-serializinghtml=1.1.4=py_0
sqlalchemy=1.3.13=py37hfa6e2cd_0
sqlite=3.33.0=h2a8f88b_0
sqlparse=0.4.1=pyh9f0ad1d_0
tabulate=0.8.7=pyh9f0ad1d_0
terminado=0.9.1=py37hc8dfbb8_1
testpath=0.4.4=py_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tifffile=2020.10.1=py_0
tk=8.6.10=he774522_1
toolz=0.11.1=py_0
tornado=6.0.4=py37h4ab8f01_2
tqdm=4.50.2=pyh9f0ad1d_0
traitlets=5.0.5=py_0
urllib3=1.25.11=py_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_3
waitress=1.4.3=py_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
websocket-client=0.57.0=py37hc8dfbb8_3
werkzeug=1.0.1=pyh9f0ad1d_0
wheel=0.35.1=py_0
widgetsnbextension=3.5.1=py37hc8dfbb8_4
win_inet_pton=1.1.0=py37hc8dfbb8_1
wincertstore=0.2=py37_0
winpty=0.4.3=4
xz=5.2.5=h62dcd97_1
yaml=0.2.5=he774522_0
zeromq=4.3.2=ha925a31_4
zfp=0.5.5=ha925a31_4
zipp=3.3.1=py_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.5=h1f3a1b7_2
